Pop expr_stack when we see a closing label
Void calls currently stay on the expression stack until a mnemonic
occurs which is not a set_local.

This doesn't work if you call a void function which is followed by a
label; in that case, the expression stack won't be popped until the next
mnemonic, which will actually move the call past the label.

Now we check the expr_stack before closing a label.