Add tests to wasmate.py
I copied and modified the test runner from the spec repo.

The collection of tests is copied from the output of running llc on the
LLVM WebAssembly CodeGen tests. For now, we'll check in this output so
everyone doesn't have to build LLVM to run these tests. In the future,
we may use a prebuilt version of LLVM for this.